+++ src/selfextract.nim
@@ -260,9 +260,6 @@ proc testConfigFile(newCon4m: string,
                addConfLoad(sbomConfName,      toStream(sbomConfig)).
                addConfLoad(sastConfName,      toStream(sastConfig)).
                addConfLoad(secretsConfName,   toStream(secretsConfig)).
-               # TODO for Theo: load the internal config file for rules
-               addConfLoad(linguistConfName,  toStream(linguistConfig)).
-               addConfLoad(techStackConfName, toStream(techStackConfig)).
                addConfLoad(coConfName,        toStream(coConfig))
 
   try:
@@ -310,8 +307,6 @@ const nocache = [getoptConfName,
                  ioConfName,
                  attestConfName,
                  defCfgFname,
-                 linguistConfName,
-                 techStackConfName,
                  coConfName,
                  embeddedConfName]